Specify size for high part. Corrects errors caught in dulpf unit tests.

+++ dulp.c
@@ -19,7 +19,7 @@ dulp(double x, double y)
     uint64_t vx = dulpval(x);
     uint64_t vy = dulpval(y);
     int lo = (vy & 1) - (vx & 1);
-    long long int hi = (vy >> 1) - (vx >> 1);
+    int64_t hi = (vy >> 1) - (vx >> 1);
     return 2.*hi + lo;
 }
 
@@ -38,6 +38,6 @@ dulpf(float x, float y)
     uint32_t vx = dulpvalf(x);
     uint32_t vy = dulpvalf(y);
     int lo = (vy & 1) - (vx & 1);
-    long int hi = (vy >> 1) - (vx >> 1);
+    int32_t hi = (vy >> 1) - (vx >> 1);
     return 2.f*hi + lo;
 }
